agent/smartscribe/pyproject.toml
[project]
name = "smartscribe"
version = "0.1.0"
description = ""
authors = [
    {name = "Abinand P",email = "abinand0911@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "copilotkit (>=0.1.34,<0.2.0)",
    "langchain-groq (>=0.2.3,<0.3.0)",
    "langgraph (>=0.2.68,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
